feat(126196): Refatora index receitas previstas
