+++ src/componentes/escolas/Paa/ElaboracaoPaa/ElaborarNovoPlano/ReceitasPrevistas/TableReceitasPrevistasPdde.js
@@ -1,18 +1,20 @@
-import React, { useState } from 'react';
+import React from "react";
 import { useGetCategoriasPddeTotais } from "./hooks/useGetCategoriasPddeTotais";
 import { Spin } from "antd";
 import { Column } from "primereact/column";
 import { DataTable } from "primereact/datatable";
 import { IconButton } from "../../../../../Globais/UI";
-import { formatMoneyBRL } from '../../../../../../utils/money';
+import { formatMoneyBRL } from "../../../../../../utils/money";
 
-const TableReceitasPrevistasPdde = ({activeTab, tabs, setActiveTab}) => {
-  const [modalForm, setModalForm] = useState({ open: false, data: null });
+const TableReceitasPrevistasPdde = ({ activeTab, tabs, setActiveTab }) => {
   const { categorias, total, isLoading } = useGetCategoriasPddeTotais();
 
   const nomeTemplate = (rowData, column) => {
     return (
-      <span style={{ color: "#00585E" }} className="font-weight-bold">
+      <span
+        style={{ color: "rgba(81, 81, 207, 1)" }}
+        className="font-weight-bold"
+      >
         {rowData.nome}
       </span>
     );
@@ -20,14 +22,14 @@ const TableReceitasPrevistasPdde = ({activeTab, tabs, setActiveTab}) => {
 
   const moneyTemplate = (field) => (rowData) => {
     return (
-      <div style={{ textAlign: 'right' }}>
+      <div style={{ textAlign: "right" }}>
         {formatMoneyBRL(rowData[field] || 0)}
       </div>
     );
   };
 
   const handleChangeTab = () => {
-    setActiveTab('detalhamento-das-acoes-pdde');
+    setActiveTab("detalhamento-das-acoes-pdde");
   };
 
   const rowClassName = (rowData) => {
@@ -45,7 +47,7 @@ const TableReceitasPrevistasPdde = ({activeTab, tabs, setActiveTab}) => {
         aria-label="Editar"
         onClick={() => handleChangeTab()}
       />
-    )
+    );
   };
 
   const totalRow = {
@@ -54,46 +56,46 @@ const TableReceitasPrevistasPdde = ({activeTab, tabs, setActiveTab}) => {
     total_valor_capital: total?.total_valor_capital || 0,
     total_valor_livre_aplicacao: total?.total_valor_livre_aplicacao || 0,
     total: total?.total || 0,
-    fixed: true
+    fixed: true,
   };
 
   return (
     <>
-    {tabs.length > 0 && activeTab === tabs[0].id ? (
-      <Spin spinning={isLoading}>
-        <DataTable
-          className="tabela-receitas-previstas mt-4"
-          value={[...categorias, totalRow]}
-          rowClassName={rowClassName}
-        >
-          <Column field="nome" header="Recursos" body={nomeTemplate} />
-          <Column
-            field="total_valor_custeio"
-            header="Custeio (R$)"
-            body={moneyTemplate('total_valor_custeio')}
-            align="right"
-          />
-          <Column
-            field="total_valor_capital"
-            header="Capital (R$)"
-            body={moneyTemplate('total_valor_capital')}
-            align="right"
-          />
-          <Column
-            field="total_valor_livre_aplicacao"
-            header="Livre Aplicação (R$)"
-            body={moneyTemplate('total_valor_livre_aplicacao')}
-            align="right"
-          />
-          <Column 
-            field="total" 
-            header="Total (R$)" 
-            body={moneyTemplate('total')}
-            align="right"
-          />
-          <Column field="acoes" header="Ações" body={acoesTemplate} />
-        </DataTable>
-      </Spin>
+      {tabs.length > 0 && activeTab === tabs[0].id ? (
+        <Spin spinning={isLoading}>
+          <DataTable
+            className="tabela-receitas-previstas mt-4"
+            value={[...categorias, totalRow]}
+            rowClassName={rowClassName}
+          >
+            <Column field="nome" header="Recursos" body={nomeTemplate} />
+            <Column
+              field="total_valor_custeio"
+              header="Custeio (R$)"
+              body={moneyTemplate("total_valor_custeio")}
+              align="right"
+            />
+            <Column
+              field="total_valor_capital"
+              header="Capital (R$)"
+              body={moneyTemplate("total_valor_capital")}
+              align="right"
+            />
+            <Column
+              field="total_valor_livre_aplicacao"
+              header="Livre Aplicação (R$)"
+              body={moneyTemplate("total_valor_livre_aplicacao")}
+              align="right"
+            />
+            <Column
+              field="total"
+              header="Total (R$)"
+              body={moneyTemplate("total")}
+              align="right"
+            />
+            <Column field="acoes" header="Ações" body={acoesTemplate} />
+          </DataTable>
+        </Spin>
       ) : null}
     </>
   );
